+++ editor/src/ek/editor/assets/editor_project.cpp
@@ -19,7 +19,7 @@ namespace ek {
 
 void editor_project_t::update_scale_factor(float scale_factor_) {
     const int uid = math::clamp(static_cast<int>(std::ceil(scale_factor_)), 1, 4);
-    scale_factor = scale_factor_;
+    scale_factor = math::clamp(scale_factor_, 1.0f, 4.0f);
     if (scale_uid != uid) {
         scale_uid = uid;
         // todo: maybe better naming `update`?
@@ -42,11 +42,17 @@ void editor_project_t::populate(bool auto_load) {
     }
 }
 
-std::string get_asset_xml_type(const path_t& path) {
-    std::string result;
+struct AssetInfoHeader {
+    std::string type;
+    bool dev = false;
+};
+
+AssetInfoHeader openAssetInfoHeader(const path_t& path) {
+    AssetInfoHeader result{};
     pugi::xml_document doc{};
     if (doc.load_file(path.c_str())) {
-        result = doc.first_child().attribute("type").as_string();
+        result.type = doc.first_child().attribute("type").as_string();
+        result.dev = doc.first_child().attribute("dev").as_bool(false);
     } else {
         EK_ERROR("XML parsing error %s", path.c_str());
     }
@@ -54,15 +60,19 @@ std::string get_asset_xml_type(const path_t& path) {
 }
 
 void editor_project_t::add_file(const path_t& path) {
-    const auto asset_type = get_asset_xml_type(base_path / path);
-    auto factory_method = type_factory[asset_type];
+    const auto header = openAssetInfoHeader(base_path / path);
+    if (header.dev && !devMode) {
+        // skip dev files for build without editor
+        return;
+    }
+    auto factory_method = type_factory[header.type];
     if (factory_method) {
         auto* asset = factory_method(path);
         assert(asset);
         asset->project = this;
         assets.push_back(asset);
     } else {
-        EK_ERROR("Editor asset type %s not found", asset_type.c_str());
+        EK_ERROR("Editor asset type '%s' not found", header.type.c_str());
     }
 }