+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- merge `name_t` and `node_state_t` to `node_t`
+- updata ImGui v1.80
+- export filters to text format layers 
+- fix macOS backspace/delete key codes
 - [audio] change api to Sound and Music instances 
 - [audio] change `simple_audio_manager` to `AudioManager`
 - [audio] fix audio loading for EDITOR mode